feat(api-client): updates collection overall layout
