+++ packages/api-client/src/components/ViewLayout/ViewLayoutSection.vue
@@ -1,9 +1,17 @@
+<script setup lang="ts">
+import { useBindCx } from '@scalar/components'
+
+const { cx } = useBindCx()
+</script>
+
 <template>
   <section
-    class="flex xl:h-full xl:min-w-0 flex-1 flex-col xl:custom-scroll bg-b-1">
+    v-bind="
+      cx('xl:custom-scroll bg-b-1 flex flex-1 flex-col xl:h-full xl:min-w-0')
+    ">
     <div
       v-if="$slots.title"
-      class="border-b min-h-11 -mb-0.25 flex items-center px-2.5 text-sm font-medium sticky top-0 bg-b-1 xl:rounded-none z-1">
+      class="-mb-0.25 bg-b-1 z-1 sticky top-0 flex min-h-11 items-center border-b px-2.5 text-sm font-medium xl:rounded-none">
       <slot name="title" />
     </div>
     <slot />
