+++ packages/api-client/src/views/Collection/Collection.vue
@@ -1,15 +1,20 @@
 <script setup lang="ts">
-import { watch } from 'vue'
+import { useScroll } from '@vueuse/core'
+import { computed, ref, watch } from 'vue'
 import { RouterView, useRouter } from 'vue-router'
 
 import ViewLayout from '@/components/ViewLayout/ViewLayout.vue'
-import ViewLayoutContent from '@/components/ViewLayout/ViewLayoutContent.vue'
+import ViewLayoutSection from '@/components/ViewLayout/ViewLayoutSection.vue'
 import { PathId } from '@/router'
 import { useActiveEntities } from '@/store'
 import CollectionNavigation from '@/views/Collection/CollectionNavigation.vue'
 
 const { activeCollection } = useActiveEntities()
 const router = useRouter()
+
+const el = ref<HTMLElement | null>(null)
+const { y } = useScroll(el)
+const isSticky = computed(() => y.value > 104)
 
 // If collection is called 'Drafts', redirect to the first request
 watch(
@@ -30,10 +35,14 @@ watch(
 </script>
 
 <template>
-  <ViewLayout class="flex divide-y md:flex-col">
-    <CollectionNavigation />
-    <ViewLayoutContent class="flex-1">
-      <RouterView />
-    </ViewLayoutContent>
+  <ViewLayout
+    ref="el"
+    class="h-fit overflow-auto pb-6 xl:overflow-auto">
+    <ViewLayoutSection class="xl:h-fit">
+      <CollectionNavigation :isSticky="isSticky" />
+      <div class="w-full md:mx-auto md:max-w-[720px]">
+        <RouterView />
+      </div>
+    </ViewLayoutSection>
   </ViewLayout>
 </template>
